switch

+++ console/src/main/java/com/ctrip/framework/drc/console/controller/v2/DrcBuildControllerV2.java
@@ -1,5 +1,6 @@
 package com.ctrip.framework.drc.console.controller.v2;
 
+import com.ctrip.framework.drc.console.config.DefaultConsoleConfig;
 import com.ctrip.framework.drc.console.param.v2.*;
 import com.ctrip.framework.drc.console.service.v2.DrcBuildServiceV2;
 import com.ctrip.framework.drc.console.vo.v2.ColumnsConfigView;
@@ -22,8 +23,14 @@ public class DrcBuildControllerV2 {
     @Autowired
     private DrcBuildServiceV2 drcBuildServiceV2;
 
+    @Autowired
+    private DefaultConsoleConfig defaultConsoleConfig;
+
     @PostMapping("mha")
     public ApiResult<Boolean> buildMha(@RequestBody DrcMhaBuildParam param) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.buildMha(param);
             return ApiResult.getSuccessInstance(true);
@@ -34,6 +41,9 @@ public class DrcBuildControllerV2 {
 
     @PostMapping("")
     public String buildDrc(@RequestBody DrcBuildParam param) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return "not allowed";
+        }
         try {
             return drcBuildServiceV2.buildDrc(param);
         } catch (Exception e) {
@@ -61,6 +71,9 @@ public class DrcBuildControllerV2 {
 
     @PostMapping("dbReplication")
     public ApiResult<List<Long>> configureDbReplications(@RequestBody DbReplicationBuildParam param) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             return ApiResult.getSuccessInstance(drcBuildServiceV2.configureDbReplications(param));
         } catch (Exception e) {
@@ -79,6 +92,9 @@ public class DrcBuildControllerV2 {
 
     @DeleteMapping("dbReplication")
     public ApiResult<Boolean> deleteDbReplications(@RequestParam long dbReplicationId) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.deleteDbReplications(dbReplicationId);
             return ApiResult.getSuccessInstance(true);
@@ -89,6 +105,9 @@ public class DrcBuildControllerV2 {
 
     @PostMapping("columnsFilter")
     public ApiResult<Boolean> buildColumnsFilter(@RequestBody ColumnsFilterCreateParam param) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.buildColumnsFilter(param);
             return ApiResult.getSuccessInstance(true);
@@ -108,6 +127,9 @@ public class DrcBuildControllerV2 {
 
     @DeleteMapping("columnsFilter")
     public ApiResult<Boolean> deleteColumnsFilter(@RequestBody List<Long> dbReplicationIds) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.deleteColumnsFilter(dbReplicationIds);
             return ApiResult.getSuccessInstance(true);
@@ -127,6 +149,9 @@ public class DrcBuildControllerV2 {
 
     @PostMapping("rowsFilter")
     public ApiResult<Boolean> buildRowsFilter(@RequestBody RowsFilterCreateParam param) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.buildRowsFilter(param);
             return ApiResult.getSuccessInstance(true);
@@ -137,6 +162,9 @@ public class DrcBuildControllerV2 {
 
     @DeleteMapping("rowsFilter")
     public ApiResult<Boolean> deleteRowsFilter(@RequestBody List<Long> dbReplicationIds) {
+        if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+            return ApiResult.getFailInstance(false, "not allowed");
+        }
         try {
             drcBuildServiceV2.deleteRowsFilter(dbReplicationIds);
             return ApiResult.getSuccessInstance(true);
